Fixed compile bug and turned Coords and PanCoords from int to double

+++ CustomCADs.App/Models/Cads/View/CadViewModel.cs
@@ -10,8 +10,8 @@
         public string Category { get; set; } = null!;
         public string CreationDate { get; set; } = null!;
         public decimal Price { get; set; }
-        public int[] Coords { get; set; } = new int[3];
-        public int[] PanCoords { get; set; } = new int[3];
+        public double[] Coords { get; set; } = new double[3];
+        public double[] PanCoords { get; set; } = new double[3];
         public short Fov { get; set; } = 90;
     }
 }
